move fork classes for fork.py
